+++ src/gevent/libuv/loop.py
@@ -37,10 +37,18 @@ class _Callbacks(AbstractCallbacks):
     def python_timer0_callback(self, watcher_ptr):
         return self.python_prepare_callback(watcher_ptr)
 
+    def python_queue_callback(self, watcher_ptr, revents):
+        watcher_handle = watcher_ptr.data
+        the_watcher = self.from_handle(watcher_handle)
+
+        the_watcher.loop._queue_callback(watcher_ptr, revents)
+
+
 _callbacks = assign_standard_callbacks(
     ffi, libuv, _Callbacks,
     [('python_sigchld_callback', None),
-     ('python_timer0_callback', None)])
+     ('python_timer0_callback', None),
+     ('python_queue_callback', None)])
 
 from gevent._ffi.loop import EVENTS
 GEVENT_CORE_EVENTS = EVENTS # export
@@ -92,6 +100,10 @@ class loop(AbstractLoop):
         self._fork_watchers = set()
         self._pid = os.getpid()
         self._default = self._ptr == libuv.uv_default_loop()
+        self._queued_callbacks = []
+
+    def _queue_callback(self, watcher_ptr, revents):
+        self._queued_callbacks.append((watcher_ptr, revents))
 
     def _init_loop(self, flags, default):
         if default is None:
@@ -128,7 +140,7 @@ class loop(AbstractLoop):
         # and call into its check and prepare handlers.
         # Note that this basically forces us into a busy-loop
         # XXX: As predicted, using an idle watcher causes our process
-        # to eat 100% CPU time. We instead use a timer with a max of a 1 second
+        # to eat 100% CPU time. We instead use a timer with a max of a .3 second
         # delay to notice signals. Note that this timeout also implements fork
         # watchers, effectively.
 
@@ -152,6 +164,30 @@ class loop(AbstractLoop):
             self._pid = curpid
             for watcher in self._fork_watchers:
                 watcher._on_fork()
+
+
+        # The contents of queued_callbacks at this point should be timers
+        # that expired when the loop began along with any idle watchers.
+        # We need to run them so that any manual callbacks they want to schedule
+        # get added to the list and ran next before we go on to poll for IO.
+        # This is critical for libuv on linux: closing a socket schedules some manual
+        # callbacks to actually stop the watcher; if those don't run before
+        # we poll for IO, then libuv can abort the process for the closed file descriptor.
+
+        # XXX: There's still a race condition here because we may not run *all* the manual
+        # callbacks. We need a way to prioritize those.
+
+        # Running these before the manual callbacks lead to some
+        # random test failures. In test__event.TestEvent_SetThenClear
+        # we would get a LoopExit sometimes. The problem occurred when
+        # a timer expired on entering the first loop; we would process
+        # it there, and then process the callback that it created
+        # below, leaving nothing for the loop to do. Having the
+        # self.run() manually process manual callbacks before
+        # continuing solves the problem. (But we must still run callbacks
+        # here again.)
+        self._prepare_ran_callbacks = self.__run_queued_callbacks()
+
         super(loop, self)._run_callbacks()
 
     def _init_and_start_prepare(self):
@@ -324,15 +360,9 @@ class loop(AbstractLoop):
         """
         Return all the handles that are open and their ref status.
         """
-
-        # XXX: Disabled because, at least on Windows, the times this
-        # gets called often produce `SystemError: ffi.from_handle():
-        # dead or bogus handle object`, and sometimes that crashes the process.
-        return []
-
-    def _really_debug(self):
         handle_state = namedtuple("HandleState",
                                   ['handle',
+                                   'type',
                                    'watcher',
                                    'ref',
                                    'active',
@@ -347,6 +377,7 @@ class loop(AbstractLoop):
             else:
                 watcher = None
             handles.append(handle_state(handle,
+                                        ffi.string(libuv.uv_handle_type_name(handle.type)),
                                         watcher,
                                         libuv.uv_has_ref(handle),
                                         libuv.uv_is_active(handle),
@@ -392,6 +423,33 @@ class loop(AbstractLoop):
         # In 1.12, the uv_loop_fork function was added (by gevent!)
         libuv.uv_loop_fork(self._ptr)
 
+    _prepare_ran_callbacks = False
+
+    def __run_queued_callbacks(self):
+        if not self._queued_callbacks:
+            return False
+
+        cbs = list(self._queued_callbacks)
+        self._queued_callbacks = []
+
+        for watcher_ptr, arg in cbs:
+            handle = watcher_ptr.data
+            if not handle:
+                # It's been stopped and possibly closed
+                assert not libuv.uv_is_active(watcher_ptr)
+                continue
+            val = _callbacks.python_callback(handle, arg)
+            if val == -1:
+                _callbacks.python_handle_error(handle, arg)
+            elif val == 1:
+                if not libuv.uv_is_active(watcher_ptr):
+                    if watcher_ptr.data != handle:
+                        if watcher_ptr.data:
+                            _callbacks.python_stop(None)
+                    else:
+                        _callbacks.python_stop(handle)
+        return True
+
 
     def run(self, nowait=False, once=False):
         # we can only respect one flag or the other.
@@ -402,17 +460,33 @@ class loop(AbstractLoop):
         if nowait:
             mode = libuv.UV_RUN_NOWAIT
 
-        # if mode == libuv.UV_RUN_DEFAULT:
-        #     print("looping in python")
-        #     ptr = self._ptr
-        #     ran_error = 0
-        #     while ran_error == 0:
-        #         ran_error = libuv.uv_run(ptr, libuv.UV_RUN_ONCE)
-        #     if ran_error != 0:
-        #         print("Error running loop", libuv.uv_err_name(ran_error),
-        #               libuv.uv_strerror(ran_error))
-        #     return ran_error
-        return libuv.uv_run(self._ptr, mode)
+        if mode == libuv.UV_RUN_DEFAULT:
+            while self._ptr and self._ptr.data:
+                # This is here to better preserve order guarantees. See _run_callbacks
+                # for details.
+                # It may get run again from the prepare watcher, so potentially we
+                # could take twice as long as the switch interval.
+                self._run_callbacks()
+                self._prepare_ran_callbacks = False
+                ran_status = libuv.uv_run(self._ptr, libuv.UV_RUN_ONCE)
+                # Note that we run queued callbacks when the prepare watcher runs,
+                # thus accounting for timers that expired before polling for IO,
+                # and idle watchers. This next call should get IO callbacks and
+                # callbacks from timers that expired *after* polling for IO.
+                ran_callbacks = self.__run_queued_callbacks()
+
+                if not ran_status and not ran_callbacks and not self._prepare_ran_callbacks:
+                    # A return of 0 means there are no referenced and
+                    # active handles. The loop is over.
+                    # If we didn't run any callbacks, then we couldn't schedule
+                    # anything to switch in the future, so there's no point
+                    # running again.
+                    return ran_status
+            return 0 # Somebody closed the loop
+
+        result = libuv.uv_run(self._ptr, mode)
+        self.__run_queued_callbacks()
+        return result
 
     def now(self):
         # libuv's now is expressed as an integer number of
@@ -515,9 +589,12 @@ class loop(AbstractLoop):
 
         return io_watcher.multiplex(events)
 
-    def timer(self, after, repeat=0.0, ref=True, priority=None):
-        if after <= 0 and repeat <= 0:
-            # Make sure we can spin the loop. See timer.
-            # XXX: Note that this doesn't have a `again` method.
-            return self._watchers.OneShotCheck(self, ref, priority)
-        return super(loop, self).timer(after, repeat, ref, priority)
+    def prepare(self, ref=True, priority=None):
+        # We run arbitrary code in python_prepare_callback. That could switch
+        # greenlets. If it does that while also manipulating the active prepare
+        # watchers, we could corrupt the process state, since the prepare watcher
+        # queue is iterated on the stack (on unix). We could workaround this by implementing
+        # prepare watchers in pure Python.
+        # See https://github.com/gevent/gevent/issues/1126
+        raise TypeError("prepare watchers are not currently supported in libuv. "
+                        "If you need them, please contact the maintainers.")